misc/HtmlTestSuite/results/test-1191.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
    <script src="support/testcases.sub.js"/>
</head>

<body>
<script>
    function buildNode(element_name, markup) {
      const e = document.createElement(element_name);
      e.innerHTML = markup;
      return e;
    }

    function toString(node) {
      const e = document.createElement("div");
      e.append(node.cloneNode(true));
      return e.innerHTML;
    }

    function assert_node_equals(node1, node2) {
      assert_equals(node1 instanceof Node, node2 instanceof Node);
      if (!(node1 instanceof Node)) return;

      node1.normalize();
      node2.normalize();
      assert_true(node1.isEqualNode(node2),
          `Node[${toString(node1)}] == Node[${toString(node2)}]`);
      if (node1 instanceof HTMLTemplateElement) {
        assert_node_equals(node1.content, node2.content);
      }
    }

    test(t =&gt; {
      let s = new Sanitizer();
      assert_throws_js(TypeError, _ =&gt; s.sanitizeFor());
      assert_throws_js(TypeError, _ =&gt; s.sanitizeFor(null));
    }, "Sanitizer.sanitizeFor() should throw.");

    test(t =&gt; {
      let s = new Sanitizer();
      assert_throws_js(TypeError, _ =&gt; s.sanitizeFor("xxx"));
    }, "Sanitizer.sanitizeFor() with one argument should throw.");

    for (const context of ["script", "iframe", "object", "div"]) {
      const should_fail = context != "div";
      test(t =&gt; {
        let result = new Sanitizer().sanitizeFor(context, "&lt;div&gt;Hello!&lt;/div&gt;");
        if (should_fail) {
          assert_equals(null, result);
        } else {
          assert_true(result instanceof HTMLElement);
        }
      }, `Sanitizer.sanitizeFor("${context}", ...) should ${should_fail ? "fail" : "pass"}.`);
    }

    async_test(t =&gt; {
      let s = new Sanitizer();
      s.sanitizeFor("div", "&lt;img src='https://bla/'&gt;");
      t.step_timeout(_ =&gt; {
        assert_equals(performance.getEntriesByName("https://bla/").length, 0);
        t.done();
      }, 1000);
    }, "Sanitizer.sanitizeFor function shouldn't load the image.");

    test(t =&gt; {
      const probe = `&lt;a href="about:blank"&gt;hello&lt;/a&gt;&lt;script&gt;con` +
          `sole.log("world!");&lt;` + `/script&gt;`;
      const expected = `&lt;a href="about:blank"&gt;hello&lt;/a&gt;`;
      for (const element of ["div", "template", "span", "table", "td",
                             "pumuckl", "custom-element", "linearGradient",
                             "svg", "svg:img", "svg:linearGradient"]) {
        assert_node_equals(
            buildNode(element, expected),
            new Sanitizer().sanitizeFor(element, probe));
      }
    }, `Sanitizer.sanitizeFor(element, ..)`);

    for (const context of ["div", "template", "table"]) {
      for (const probe of ["&lt;em&gt;Hello&lt;/em&gt;", "&lt;td&gt;data&lt;/td&gt;"]) {
        test(t =&gt; {
          assert_node_equals(
              buildNode(context, probe),
              new Sanitizer().sanitizeFor(context, probe));
        }, `Sanitizer.sanitizeFor("${context}", "${probe}") obeys parse context.`);
      }
    }

    for (const testcase of testcases) {
      test(t =&gt; {
        let s = new Sanitizer(testcase.config_input);
        assert_node_equals(
            buildNode("template", testcase.result),
            s.sanitizeFor("template", testcase.value));
      }, "Sanitizer.sanitizeFor with config: " + testcase.message);
    }
</script>
</body>
</html>
